Fixes Singleton bug
its gross, but it works.

Basically calls a singleton to delay the findcomponent call from the relevant scripts, as the root of the issue was that the non singleton objects were grabbing the version that was going to be destroyed instead of the singleton one.

Maybe a way to improve is storing the singletons in the game manager, and then having the scripts just poll the game manager for the objects? irrelevant for the moment